Split util.py into examples.py and parser.py.
