+++ spacy_llm/tasks/lemma/parser.py
@@ -1,25 +1,14 @@
-from typing import Dict, Iterable, List
+from typing import Iterable
 
 from spacy.tokens import Doc
-from spacy.training import Example
-
-from ...ty import FewshotExample
-from ..templates import read_template
-
-DEFAULT_LEMMA_TEMPLATE_V1 = read_template("lemma")
-
-
-class LemmaExample(FewshotExample):
-    text: str
-    lemmas: List[Dict[str, str]]
-
-    @classmethod
-    def generate(cls, example: Example) -> "LemmaExample":
-        lemma_dict = [{t.text: t.lemma_} for t in example.reference]
-        return LemmaExample(text=example.reference.text, lemmas=lemma_dict)
 
 
 def parse_responses_v1(docs: Iterable[Doc], responses: Iterable[str]) -> Iterable[Doc]:
+    """Parses LLM responses for spacy.lemma.v1 and maps them onto Docs.
+    docs (Iterable[Doc]): Docs to map responses to.
+    responses (Iterable[str]): LLM responses.
+    RETURNS (Iterable[Doc]): Updated docs with mapped LLM responses.
+    """
     for doc, prompt_response in zip(docs, responses):
         parsed_response = [
             [pr_part.strip() for pr_part in pr.split(":")]
